% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{add_arrows}
\alias{add_arrows}
\title{Add arrows to a swimmer plot}
\usage{
add_arrows(
  data = NULL,
  mapping = NULL,
  position = "identity",
  arrow = NULL,
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed"
)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{\link[=ggswim]{ggswim()}}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}

\item{arrow}{A column indicating what swim lanes should have arrows applied.
The column must be a logical data type (T/F).}

\item{arrow_colour}{Border/line color to use for the arrow. Default "black".}

\item{arrow_head_length}{A unit specifying the length of the arrow head (from tip to base).
Must be a ggplot2 \code{unit()} object. Default \code{ggplot2::unit(0.25, "inches")}.}

\item{arrow_neck_length}{The length of the neck of the arrow from the end of a
swim lane to the base of the arrow head. Either an integer value or a column
specifier. The default, \code{NULL}, sets a value proportional to the max lane value.}

\item{arrow_fill}{Fill color/colour to use for the arrow head (if closed). Default \code{NULL}.}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head should
be a closed triangle. Default "closed."}

\item{...}{Other arguments passed to \code{ggswim()}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\description{
Add arrows to the ends of swimmer plot lanes to indicate unknown statuses
or continued record-level trajectories.
}
\details{
\code{add_arrows()} wraps a new \code{\link[=geom_segment]{geom_segment()}} layer by adding a zero-length
segment at the right end of swimmer lanes. This approach allows users to
specify \code{arrow_neck_length} which can be useful for tracking and visualizaing
time in between markers
}
\examples{
add_arrows(data = patient_status,
           mapping = aes(xend = end_time, y = pt_id),
           arrow = arrow,
           arrow_neck_length = time_from_today,
           arrow_colour = "forestgreen",
           arrow_fill = "forestgreen")

}
