% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.R
\name{add_columns}
\alias{add_columns}
\title{Add empty columns based on supplied day_vector}
\usage{
add_columns(df, columns)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a dataframe
}
\description{
Add empty columns to a dataframe based on supplied day_vector
}
\keyword{internal}
