% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker_new.R
\name{add_marker_new}
\alias{add_marker_new}
\title{add_marker title}
\usage{
add_marker_new(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{data where markers reside, if in ggswim object leave \code{NULL} (default)}

\item{...}{Other arguments passed on to layer(). These are often aesthetics,
used to set an aesthetic to a fixed value, like colour = "red" or size = 3.
They may also be parameters to the paired geom/stat.}
}
\value{
A ggswim object
}
\description{
A short description...
}
