% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{append_guide_capture}
\alias{append_guide_capture}
\title{Append guide_capture list elements}
\usage{
append_guide_capture(ggswim_obj, current_layer, name)
}
\arguments{
\item{ggswim_obj}{ggswim_obj}

\item{current_layer}{the current ggplot layer}

\item{name}{name of the current layer, defined by the "name" argyument in \code{add_marker()}}
}
\value{
a ggswim object
}
\description{
Internal utility function that appends sub lists onto the \code{guide_capture} list
of a \code{ggswim} object
}
\keyword{internal}
