% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{apply_geom_points}
\alias{apply_geom_points}
\title{apply an arbitrary number of \code{geom_point()}'s to a ggplot}
\usage{
apply_geom_points(plot, markers, id)
}
\arguments{
\item{plot}{a ggplot2 object}

\item{markers}{a vector of columns that will comprise point markers on the
swimmer plot}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}
}
\value{
a ggplot object
}
\description{
This function applies any number of necessary \code{geom_point()}
objects to a \code{ggplot2} chain as declared in \code{markers}.
}
\keyword{internal}
