% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{apply_guide_overrides}
\alias{apply_guide_overrides}
\title{Assign guide overrides}
\usage{
apply_guide_overrides(ggswim_obj, ...)
}
\arguments{
\item{ggswim_obj}{description}

\item{...}{Additional arguments from the parent function}
}
\value{
a ggswim object
}
\description{
Helper function for assigning guide override values to a ggplot function for
use with \code{add_markers}
}
\keyword{internal}
