% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_updated_legend_order}
\alias{apply_updated_legend_order}
\title{Apply updated ggplot legend order}
\usage{
apply_updated_legend_order(gg, lanes, markers)
}
\arguments{
\item{gg}{A \code{ggplot} object}

\item{lanes}{A list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{markers}{a named list defining marker events on a \code{lane} in either
numeric shape or emoji form}
}
\value{
a ggplot object
}
\description{
This helper function takes the updated legend order from \code{get_gg_legend_order}
and applies directly to the labels of a ggplot object's legend.
}
\keyword{internal}
