% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.R
\name{as_swim_tbl}
\alias{as_swim_tbl}
\title{Add swim_tbl S3 class}
\usage{
as_swim_tbl(x)
}
\arguments{
\item{x}{an object to class}
}
\value{
The object with \code{swim_tbl} S3 class
}
\description{
Add swim_tbl S3 class
}
\keyword{internal}
