% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{bind_layer_data}
\alias{bind_layer_data}
\title{Bind layer dataframes for legend}
\usage{
bind_layer_data(ggswim_obj, layer_indices, fixed_colours = NULL)
}
\arguments{
\item{ggswim_obj}{A ggswim object}

\item{layer_indices}{Layer indexes}

\item{fixed_colours}{Dataframe of fixed colour callouts}
}
\value{
A dataframe
}
\description{
Internal helper function that returns layer data from \code{get_layer_data()}
as a bound dataframe to help with legend guide definitions.
}
\keyword{internal}
