% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{bootstrap}
\alias{bootstrap}
\title{Retrieve Bootstrap unicode}
\usage{
bootstrap(aliases)
}
\arguments{
\item{aliases}{A string or vector of strings to retrieve}
}
\value{
Unicode text
}
\description{
When assigning Bootstrap icons as glyphs, \code{\link[=bootstrap]{bootstrap()}} should be used to
convert the alias string to the appropriate Unicode format.

All \code{aliases} should be prepended with "bs".
}
\examples{
boostrap("bs-car-front")
}
