% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{build_ggswim}
\alias{build_ggswim}
\title{Apply ggswim fixes and display updates}
\usage{
build_ggswim(ggswim_obj)
}
\arguments{
\item{ggswim_obj}{A ggswim object}
}
\value{
A ggplot2 object
}
\description{
Users should generally never have to call this function, but it is used to
invoke final changes before plot rendering and a custom print method via
\code{\link[=print.ggswim_obj]{print.ggswim_obj()}}.
}
\details{
In its current state, \code{build_ggswim()} can only work with a pre-rendered
ggswim plot object, therefore it cannot be added to the \code{+} operator chain.
}
\examples{
ggswim_obj <- ggplot2::ggplot(data = patient_data) +
  geom_swim_lane(
    mapping = aes(
      x = start_time,
      y = pt_id,
      xend = end_time,
      color = disease_assessment
    )
  )
build_ggswim(ggswim_obj)
}
