% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_error}
\alias{capture_error}
\title{Error Capturing and Testing}
\usage{
capture_error(expr)
}
\arguments{
\item{expr}{An expression to test}
}
\description{
Internal function for testing error handling in the test suite.
}
\keyword{internal}
