% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_coerced_data}
\alias{check_coerced_data}
\title{check for coerced data inside a ggswim function}
\usage{
check_coerced_data(expr)
}
\arguments{
\item{expr}{a character vector to test for existence in the names of a
dataset}
}
\description{
Supply users with an error when a coercion to data elements occurs inside of
\code{ggswim()} or \code{add_marker()} and cannot be rectified with
\code{retrieve_original_aes()}.
}
\details{
ggswim references internal ggplot layers and any aesthetic mapping
required for downstream rendering. If a user applies a coercion in a function,
ggswim may not be able to parse the original variable. For example, in
\verb{ggswim(mtcars, aes(x = hp, y = cyl, color = factor(disp)}),
\code{rlang::get_expr()} will see the color mapping aesthetic as \code{factor(disp)},
and not \code{disp}.
}
\keyword{internal}
