% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_for_na_legend_discrepancy}
\alias{check_for_na_legend_discrepancy}
\title{Check for plot legend discrepancies due to missing data}
\usage{
check_for_na_legend_discrepancy(ggswim_obj, override)
}
\arguments{
\item{ggswim_obj}{A ggswim object}

\item{override}{An internal list object responsible for legend \code{guide()} rebuilds}
}
\value{
A boolean
}
\description{
There is a known issue with ggswim \code{add_markers()} where missing data in prior
layers can cause issues when compared to non-missing data in subsequent layers.
Due to the way ggswim handles manipulation of the legend to allow for combination
of various layer types under the "colour" layer, NA values must be dropped from
the legend display if they are detected in the data but not in the \code{ggswim_obj}.

This error occurs due to a mismatch in expected values supplied to the \code{guide()}
via the \code{override} reference list.
}
\keyword{internal}
