% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ggswim_obj}
\alias{check_ggswim_obj}
\title{check build_ggswim receives a ggswim class object}
\usage{
check_ggswim_obj(obj)
}
\arguments{
\item{obj}{The ggplot object passed to the ggswim print method. Ideally, should
always have an enforced \code{ggswim_obj} class type when ggswim is being used.}
}
\description{
Supply users with an error when \code{build_ggswim()} is used outside of the
context of a \code{ggswim_obj} class object.
}
\keyword{internal}
