% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_marker_label_aes}
\alias{check_marker_label_aes}
\title{check add_markers for label arguments}
\usage{
check_marker_label_aes(mapping)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}
}
\description{
For proper display of label values in the legend, a \code{color}/\code{colour} mapping
value along with a name specifier is typically needed. It is not mandatory to
supply a naming parameter, and a warning will be displayed if labels are used
but no naming param is detected.

An error will be thrown if a label naming parameter is specified, but no
label value parameter is included. Default ggplot2 behavior is to allow this
and throw a warning, but we instead wish this to be an error.
}
\keyword{internal}
