% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_missing_aes_params}
\alias{check_missing_aes_params}
\title{check for missing aes mappingparams}
\usage{
check_missing_aes_params(mapping, params, parent_func)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{params}{a vector of params to check for in string format}

\item{parent_func}{The function in which this is being called, to be
referenced in the message output}
}
\description{
Utility check to block users from submitting functions that are missing key
parameters.
}
\details{
A key use of this check is a stopgap for missing \code{x} and \code{y} aes parameters
which should be valid but don't currently work with our setup.
}
\keyword{internal}
