% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_missing_params}
\alias{check_missing_params}
\title{check for missing params}
\usage{
check_missing_params(param, name, parent_func)
}
\arguments{
\item{param}{A parameter to check for \code{NULL}}

\item{name}{The name of the parameter checked}

\item{parent_func}{The function in which this is being called, to be
referenced in the message output}
}
\description{
Utility check to block users from submitting functions that are missing key
parameters. Tests a single parameter for being \code{NULL}.
}
\keyword{internal}
