% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_supported_mapping_aes}
\alias{check_supported_mapping_aes}
\title{check for unsupported mapping args}
\usage{
check_supported_mapping_aes(mapping, unsupported_aes, parent_func)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{unsupported_aes}{A character vector}

\item{parent_func}{The function in which this is being called, to be
referenced in the message output}
}
\description{
\code{ggswim()} and \code{add_marker()} do not support all of the aesthetic mappings
their wrapped functions do. By design, this helps \code{build_ggswim()} work
appropriately.
}
\keyword{interal}
