% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_valid_fixed_maker_scales}
\alias{check_valid_fixed_maker_scales}
\title{Fixed Marker Scale Validation}
\usage{
check_valid_fixed_maker_scales(ggswim_obj, ref_layer_info)
}
\arguments{
\item{ggswim_obj}{A ggswim object}

\item{ref_layer_info}{internal reference layer list created in \code{build_ggswim()}}
}
\description{
Check fixed markers for single scale use and for single title assignment.
}
\details{
ggswim cannot currently support fixed markers with multiple scales or titles,
this causes the legend to display incorrectly and for titles to be
misattributed. Instead, we wish to catch this early and enforce behavior.
}
\keyword{internal}
