% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_arrow.R
\docType{data}
\name{.default_arrow_colours}
\alias{.default_arrow_colours}
\alias{.default_arrow_fills}
\alias{.default_arrow_types}
\alias{.default_arrow_limits}
\title{ggswim arrow defaults}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
.default_arrow_colours

.default_arrow_fills

.default_arrow_types

.default_arrow_limits
}
\description{
ggswim arrow defaults
}
\examples{
ggswim::.default_arrow_colours
ggswim::.default_arrow_fills
ggswim::.default_arrow_types
ggswim::.default_arrow_limits

}
\keyword{datasets}
