% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_marker.R
\docType{data}
\name{.default_glyphs}
\alias{.default_glyphs}
\alias{.default_colours}
\title{ggswim marker defaults}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
.default_glyphs

.default_colours
}
\description{
ggswim marker defaults
}
\examples{
ggswim::.default_glyphs
ggswim::.default_colours

}
\keyword{datasets}
