% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{end_study_events}
\alias{end_study_events}
\title{End of Study Events dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\source{
This dataset is for demonstration purposes only.
}
\usage{
data(end_study_events)
}
\description{
These datasets provide information about patients, infusion events, and end of
study events. Trial data has been de-identified and randomized for general use.
}
\details{
\code{end_study_events} contains pre-formatted time series data related to end of
study events where patients left the trial for varying reasons. This dataset
is most applicable for use with \code{geom_swim_marker()}.
}
\examples{
end_study_events
}
\keyword{datasets}
