% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline_legend.R
\name{filter_list_elements}
\alias{filter_list_elements}
\title{Find matching list element}
\usage{
filter_list_elements(input_list, name_to_match, value_to_match)
}
\value{
a list
}
\description{
A short description...
}
\keyword{internal}
