% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_legend.R
\name{fix_legend}
\alias{fix_legend}
\title{Fix legend display for multiple layers}
\usage{
fix_legend(ggswim_obj)
}
\arguments{
\item{ggswim_obj}{A ggswim object}
}
\value{
A ggswim object
}
\description{
This function is primarily helpful when combining labels (like unicode emojis)
and point (like shapes). This function seeks to override the values and present
them in the same layer of the legend output.
}
\details{
In its current state, \code{fix_legend()} can only work with a pre-rendered ggswim
plot object, therefore it cannot be added to the \code{+} operator chain.
}
