% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swim_point.R
\docType{data}
\name{geom_swim_point}
\alias{geom_swim_point}
\alias{GeomSwimPoint}
\title{Add markers of interest swimmer plots - point}
\usage{
geom_swim_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Add markers of interest swimmer plots - point
}
\keyword{datasets}
