% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swim_lane.R
\docType{data}
\name{GeomSwimLane}
\alias{GeomSwimLane}
\title{#' @rdname geom_swim_lane
#' @format NULL
#' @usage NULL
#' @export}
\format{
An object of class \code{GeomSwimLane} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomSwimLane
}
\description{
#' @rdname geom_swim_lane
#' @format NULL
#' @usage NULL
#' @export
}
\keyword{datasets}
