% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_assigned_colors}
\alias{get_assigned_colors}
\title{Reshape lane_colors}
\usage{
get_assigned_colors(df, gg, lanes, lane_colors, markers, shape_colors)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{gg}{a \code{ggplot} object}

\item{lanes}{a list of character strings that define the colored line segments
for \code{id}. Colors are supplied by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{lane_colors}{a character vector defined in the \code{swim_tbl} that assigns
user-defined colors to \code{lanes}}

\item{markers}{a named list defining marker events on a \code{lane} in either
standard numeric ggplot 2 shapes, emoji, or unicode form .
Shapes can be supplied as character strings or integers.}

\item{shape_colors}{If providing shapes for \code{markers}, provide vector
specification for the colors those shapes should be. Default \code{NULL} for
non-shapes, default \code{ggplot2} colors if shapes with no colors specified.}
}
\value{
a named vector
}
\description{
In order to assign the correct colors to \code{ggplot2::scale_color_manual()},
it is necessary to account for the order of the legend output in combination
with the assigned \code{markers}. This way, only lanes that are present get
assigned the appropriate colors, in the order of the legend, and markers are
assigned \code{NA}.
}
\keyword{internal}
