% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_assigned_line_colors}
\alias{get_assigned_line_colors}
\title{Reshape lane_colors}
\usage{
get_assigned_line_colors(gg, lane_colors)
}
\arguments{
\item{gg}{a \code{ggplot} object}

\item{lane_colors}{a vector of character strings to assign to the \code{lanes},
written in the same order. Optional, if left \code{NULL}, default colors applied}
}
\value{
a vector
}
\description{
In order to assign the correct colors to \code{ggplot2::scale_color_manual()},
it is necessary to account for the order of the legend output in combination
with the assigned \code{markers}.
}
\keyword{internal}
