% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{get_ggswim_obj_from_env}
\alias{get_ggswim_obj_from_env}
\title{Get ggswim object from an enviornment variable}
\usage{
get_ggswim_obj_from_env(env)
}
\arguments{
\item{env}{an enviornment object}
}
\value{
A ggswim object
}
\description{
Determine the parent environment component containing a \code{ggswim_obj} class
type and return the object
}
\keyword{internal}
