% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_guide_values}
\alias{get_guide_values}
\title{Create guide values for legend}
\usage{
get_guide_values(df, gg, emoji_or_shape, lanes, markers)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{gg}{A \code{ggplot} object}

\item{emoji_or_shape}{One of "emoji" or "shape", determines whether to use
\code{geom_label()} or \code{geom_point()}}

\item{lanes}{A list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{markers}{a named list defining marker events on a \code{lane} in either
numeric shape or emoji form}
}
\value{
a variable
}
\description{
Programmatically assign and retrieve line and shape/label values to supply to
\code{ggplot2::guide()}.
}
\keyword{internal}
