% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swim_label.R
\name{get_label_override}
\alias{get_label_override}
\title{Fix Label Legend}
\usage{
get_label_override(plot, layer)
}
\arguments{
\item{plot}{The current plot}

\item{layer}{The new label layer being added}
}
\value{
A dataframe, acting as a key for \code{\link[=guide_legend]{guide_legend()}} label aes overrides
}
\description{
Fixes the legend for labels, while accounting for use of \code{\link[=new_scale_colour]{new_scale_colour()}},
so that labels appear over legend glyphs.
}
