% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_lane_colors}
\alias{get_lane_colors}
\title{Get lane colors}
\usage{
get_lane_colors(lanes, lane_colors)
}
\arguments{
\item{lanes}{A list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{lane_colors}{a character vector defined in the \code{swim_tbl} that assigns
user-defined colors to \code{lanes}}
}
\value{
A named vector
}
\description{
This internal function detects whether or not lane colors are supplied in a
\code{swim_tbl} and returns them with the appropriate \code{lanes} names. If no colors
are given (i.e. \code{NULL}), default ggplot2 colors are supplied via the \code{scales}
package.
}
