% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{get_overrides}
\alias{get_overrides}
\title{Get overrides list for \code{guides()}}
\usage{
get_overrides(ggswim_obj, label_layer_data, point_layer_data)
}
\arguments{
\item{ggswim_obj}{A ggswim object}

\item{label_layer_data}{description}

\item{point_layer_data}{description}
}
\value{
a list
}
\description{
Creates a list of override definitions to pass to \code{guides()} \code{override.aes}.
}
\keyword{internal}
