% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{get_ref_layer_info}
\alias{get_ref_layer_info}
\title{Populate reference layer information list}
\usage{
get_ref_layer_info(ggswim_obj)
}
\arguments{
\item{ggswim_obj}{A ggswim object}
}
\value{
a list
}
\description{
The reference layer information list is used for location of layer type indices
and fixed color definitions.
}
\keyword{internal}
