% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim-package.R
\docType{package}
\name{ggswim-package}
\alias{ggswim-package}
\alias{_PACKAGE}
\title{ggswim: Create Swimmer Plots with ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create swimmer plots using the ggplot2 framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chop-cgtinformatics.github.io/ggswim/}
  \item \url{https://github.com/CHOP-CGTInformatics/ggswim}
  \item Report bugs at \url{https://github.com/CHOP-CGTInformatics/ggswim/issues}
}

}
\author{
\strong{Maintainer}: Richard Hanna \email{richardshanna91@gmail.com} (\href{https://orcid.org/0009-0005-6496-8154}{ORCID})

Authors:
\itemize{
  \item Stephan Kadauke \email{kadaukes@chop.edu} (\href{https://orcid.org/0000-0003-2996-8034}{ORCID})
  \item Ezra Porter \email{ezrajporter@gmail.com} (\href{https://orcid.org/0000-0002-4690-8343}{ORCID})
}

}
\keyword{internal}
