% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim_new.R
\name{ggswim_new}
\alias{ggswim_new}
\title{ggswim title}
\usage{
ggswim_new(data, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{mapping}{description}

\item{...}{}
}
\description{
A short description...
}
