% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griddraw.R
\name{grid.draw.ggswim_obj}
\alias{grid.draw.ggswim_obj}
\title{Draw ggswim object}
\usage{
\method{grid.draw}{ggswim_obj}(x, recording = TRUE)
}
\arguments{
\item{x}{an object of class 'ggswim_obj'}

\item{recording}{A logical value to indicate whether the drawing
    operation should be recorded on the Grid display list.}
}
\value{
None
}
\description{
\code{grid::grid.draw()} methods for objects of classes 'ggswim_obj' and 'ggcuminc'.
These are implemented to allow users to directly call \code{ggplot2::ggsave()}
on 'ggswim' figures.
}
\examples{
ggswim(
  data = patient_data,
  mapping = aes(
    x = delta_t0_months,
    y = pt_id,
    fill = disease_assessment
  )
) |>
  grid.draw()
}
\keyword{internal}
