% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{infusion_events}
\alias{infusion_events}
\title{Infusion Events dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.
}
\source{
This dataset is for demonstration purposes only.
}
\usage{
data(infusion_events)
}
\description{
These datasets provide information about patients, infusion events, and end of
study events. Trial data has been de-identified and randomized for general use.
}
\details{
\code{infusion_events} contains pre-formatted time series data related to infusions
and reinfusions for patients in \code{patient_data}. This dataset is most applicable
for use with \code{add_marker()}.
}
\examples{
infusion_events
}
\keyword{datasets}
