% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_override}
\alias{insert_override}
\title{Insert override layer element}
\usage{
insert_override(data, current_layer, mapping, ignore_mapping)
}
\arguments{
\item{data}{The data responsible for the current layer}

\item{current_layer}{An integer value corresponding to the current working layer}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{ignore_mapping}{A vector of characters for mapping values to ignore.
Will typically match "required" \code{aes} elements from ggplot geom functions}
}
\value{
A ggplot layer object appended with the \code{overrides} list object
}
\description{
A helper function that wraps \code{get_layer_data} and adds the result to a ggplot
layer.
}
\keyword{internal}
