% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric_coercible}
\alias{is_numeric_coercible}
\title{Detect if object is numeric or numeric-coercible}
\usage{
is_numeric_coercible(data)
}
\arguments{
\item{data}{a vector or variable to test}
}
\value{
a boolean
}
\description{
This utility function takes a vector or variable and attempts to detect
whether it is numeric or numeric-coercible (i.e. a numeric character string).
}
\details{
At present, this is only intended to be used in detecting coercible character
strings, not factors or boolean values since coercion to numeric is both
possible and meaningful.
}
