% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{load.fa}
\alias{load.fa}
\title{Load fontawesome fonts for use with ggswim}
\usage{
load.fa(font = "fa.ttf")
}
\arguments{
\item{font}{an internal TTF file}
}
\description{
ggswim comes installed with fontawesome free solid fonts. By default, \code{\link[=load.fa]{load.fa()}}
is enabled when ggswim is imported.
}
\examples{
load.fa()

}
