% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{patient_data}
\alias{patient_data}
\alias{infusion_events}
\alias{end_study_events}
\title{Patient Datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 83 rows and 9 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.
}
\source{
This dataset is for demonstration purposes only.
}
\usage{
data(patient_data)

data(infusion_events)

data(end_study_events)
}
\description{
These datasets provide information about patients, infusion events, and end of
study events.
}
\examples{

patient_data

infusion_events

end_study_events
}
\keyword{datasets}
