% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggswim_layer}
\alias{print.ggswim_layer}
\title{Print ggswim layer object}
\usage{
\method{print}{ggswim_layer}(x, ..., flat = TRUE)
}
\arguments{
\item{x}{a ggswim layer object}

\item{...}{not used}
}
\value{
a layer object
}
\description{
Print ggswim layer object
}
\examples{
print(
  geom_swim_lane(
    data = patient_data,
    mapping = aes(
      x = start_time,
      y = pt_id,
      xend = end_time,
      color = disease_assessment
    )
  )
)
}
\keyword{internal}
