% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_original_aes}
\alias{retrieve_original_aes}
\title{Retrieve original vars from coerced vars}
\usage{
retrieve_original_aes(data, aes_mapping, aes_var)
}
\arguments{
\item{data}{the data responsible for the current layer}

\item{aes_mapping}{a list of mapping data (i.e. \code{unlist(mapping)})}

\item{aes_var}{the aesthetic variable to test for (ex: \code{color}, \code{shape})}
}
\value{
The original variable name as a character string
}
\description{
Detect instances where users manipulate \code{aes()} names that ggswim
requires to access and identify layer types downstream. In instances where
such a coercion is detected, attempt to retrieve the original \code{aes()} name.
}
\details{
ggswim references internal ggplot layers and any aesthetic mapping
required for downstream rendering. If a user applies a coercion in a function,
ggswim may not be able to parse the original variable. For example, in
\verb{ggswim(mtcars, aes(x = hp, y = cyl, color = factor(disp)}),
\code{rlang::get_expr()} will see the color mapping aesthetic as \code{factor(disp)},
and not \code{disp}.
}
\keyword{internal}
