% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{search_aliases}
\alias{search_aliases}
\title{Utility Function to Search Aliases Across Icon Libraries}
\usage{
search_aliases(str = "", type = "solid", approximate = FALSE)
}
\arguments{
\item{str}{A character string alias to search the specified icon library.
If left empty (default ""), it returns all available aliases.}

\item{type}{For FontAwesome dataset only. Specifies the subset to search within.
One of "solid", "regular", or "brands". Ignored for other datasets. Default is "solid".}

\item{approximate}{Logical. If \code{TRUE}, performs approximate matching using \code{agrep}.
If \code{FALSE} (default), performs exact matching using \code{grep}.}
}
\value{
A sorted character vector of matching aliases.
}
\description{
A generic function to search for aliases within specified icon libraries.
}
\keyword{internal}
