% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{search_fontawesome}
\alias{search_fontawesome}
\title{Search for FontAwesome aliases to include in ggswim}
\usage{
search_fontawesome(str = "", type = "solid", approximate = FALSE)
}
\arguments{
\item{str}{A character string alias to search the icon data available.
If left empty, the default "", returns all available aliases.}

\item{type}{A character string denoting which FontAwesome library to search.
One of "solid", "regular", or "brands". Default "solid".}

\item{approximate}{Use approximate or exact matching, TRUE/FALSE. Default \code{FALSE}.}
}
\value{
Matching aliases from the available FontAwesome data
}
\description{
Check strings against the available aliases for FontAwesome icons.
}
\examples{
search_fontawesome("fa-car")

}
