% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.R
\name{streamline}
\alias{streamline}
\title{Streamline data for ggswim plotting}
\usage{
streamline(df, id, time, events, reference_event, markers = NULL, lanes)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time}{the x-axis variable of the swimmer plot, typically a
function of time}

\item{events}{the column that will supply definitions for the \code{reference_event},
\code{markers}, and \code{lanes}}

\item{reference_event}{a character string found in \code{events} that establishes
the time-zero reference point for the x-axis}

\item{markers}{A character vector that will comprise point markers on the
swimmer plot. Optional, default \code{NULL}}

\item{lanes}{Columns that indicate line changes, i.e. color changes
for individual swim lanes.}
}
\value{
a swim_tbl object
}
\description{
Prepare a dataset for use with \code{ggswim()} and ensure proper elements are made
available.
}
\details{
Requirements for a dataset include an identifiable primary key column,
a time column, and any number of binary indicator columns.
}
