% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline_legend.R
\name{streamline_legend}
\alias{streamline_legend}
\title{Streamline the ggswim legend}
\usage{
streamline_legend(..., environment = parent.frame())
}
\arguments{
\item{...}{}
}
\value{
a ggswim object
}
\description{
A short description...
}
