% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_ggswim}
\alias{try_ggswim}
\title{Capture ggswim errors with custom error handling}
\usage{
try_ggswim(expr, call = caller_env())
}
\arguments{
\item{expr}{an expression for building and printing a ggswim object}

\item{call}{the calling environment to use in the warning message}
}
\value{
If successful, build and print the ggswim object. Otherwise an error.
}
\description{
Capture ggswim errors with custom error handling
}
\keyword{internal}
