% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_gg_legend_order}
\alias{update_gg_legend_order}
\title{Get ggplot legend order}
\usage{
update_gg_legend_order(gg, lanes, markers)
}
\arguments{
\item{gg}{a ggplot object}

\item{lanes}{A list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{markers}{a named list defining marker events on a \code{lane} in either
numeric shape or emoji form}
}
\value{
a character vector
}
\description{
Helper function to retrieve the existing legend order for a ggplot object
}
\keyword{internal}
