% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ggswim.R
\name{update_ggswim}
\alias{update_ggswim}
\title{Apply ggswim fixes and display updates}
\usage{
update_ggswim(ggswim_obj)
}
\arguments{
\item{ggswim_obj}{A ggswim object}
}
\value{
A ggswim object
}
\description{
This function seeks to correct the output of manual overrides introduced by
\code{add_marker()} depending on the combination of layer types the user provides.

It is to be run automatically in the background as a print method via
\code{print.ggswim_obj()}.
}
\details{
In its current state, \code{update_ggswim()} can only work with a pre-rendered
ggswim plot object, therefore it cannot be added to the \code{+} operator chain.

\code{update_ggswim()} makes use of \code{ggplot2::guides()} to dynamically override
displays in the layers of the ggswim legend. It also applies a call to
\code{ggplot2::scale_color_manual()} in applicable cases where a user calls out
a static \code{color}/\code{colour} argument in addition to the required \code{color}
mapping aesthetic (handled by arg: \code{name}).
}
