% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{update_missing_guide_overrides}
\alias{update_missing_guide_overrides}
\title{Replace NA guide values with defaults}
\usage{
update_missing_guide_overrides(ggswim_obj, ...)
}
\arguments{
\item{ggswim_obj}{description}

\item{...}{Additional arguments from the parent function}
}
\value{
a ggswim object
}
\description{
Switch NA values in \code{guide_override}s with default values
}
\keyword{internal}
