% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{add_arrows}
\alias{add_arrows}
\title{Add arrows to plot using geom_segment}
\usage{
add_arrows(
  data,
  ggswim_obj,
  mapping,
  position,
  arrow,
  arrow_colour,
  arrow_head_length,
  arrow_neck_length,
  arrow_fill,
  arrow_type
)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{ggswim_obj}{A ggswim object}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.
More information about accepted mapping arguments can be found in \strong{Aesthetics}.}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}

\item{arrow}{A column indicating what swim lanes should have arrows applied.
The column must be a logical data type (T/F).}

\item{arrow_colour}{Border/line color to use for the arrow. Default "black".}

\item{arrow_head_length}{A unit specifying the length of the arrow head (from tip to base).
Must be a ggplot2 \code{unit()} object. Default \code{ggplot2::unit(0.25, "inches")}.}

\item{arrow_neck_length}{The length of the neck of the arrow from the end of a
swim lane to the base of the arrow head. Either an integer value or a column
specifier. The default, \code{NULL}, sets a value proportional to the max lane value.}

\item{arrow_fill}{Fill color/colour to use for the arrow head (if closed). Default \code{NULL}.}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head should
be a closed triangle. Default "closed."}

\item{...}{Other arguments passed to \code{ggswim()}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\description{
This helper function is triggered when a user requests arrows to
appear in \code{ggswim()}. It uses the \code{geom_segment()} function to supply them by
Adding a 0-length segment at the end of the swim lanes and then tacking on
arrows using the \code{arrow} argument.
}
\keyword{internal}
