% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R
\name{add_arrows}
\alias{add_arrows}
\title{Add arrows to a swimmer plot}
\usage{
add_arrows(
  data = NULL,
  mapping = NULL,
  position = "identity",
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed"
)
}
\arguments{
\item{data}{A dataframe prepared for use with \code{ggswim}}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{position}{Position adjustment. ggswim accepts either "stack", or
"identity" depending on the use case. Default "identity".}

\item{arrow_colour}{The color of the arrow head}

\item{arrow_head_length}{A unit specifying the length of the arrow head
(from tip to base).}

\item{arrow_neck_length}{Value specifying neck length from end of segment
to arrow head base}

\item{arrow_fill}{The color to fill the arrow head with. Only applicable when
\code{arrow_type == "closed"}}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head
should be a closed triangle.}
}
\description{
Add arrows to the ends of swimmer plot lanes to indicate unknown statuses
or continued record-level trajectories.
}
\details{
\code{add_arrows()} wraps a new \code{\link[=geom_segment]{geom_segment()}} layer by adding a zero-length
segment at the right end of swimmer lanes. This approach allows users to
specify \code{arrow_neck_length} which can be useful for tracking and visualizing
time in between markers
}
\examples{
\dontrun{
add_arrows(
  data = patient_status,
  mapping = aes(xend = end_time, y = pt_id),
  arrow_neck_length = time_from_today,
  arrow_colour = "forestgreen",
  arrow_fill = "forestgreen"
)
}
}
