% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R
\name{add_arrows}
\alias{add_arrows}
\title{Add arrows to a swimmer plot}
\usage{
add_arrows(
  data = NULL,
  mapping = NULL,
  position = "identity",
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed"
)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{\link[=ggswim]{ggswim()}}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}
}
\description{
Add arrows to the ends of swimmer plot lanes to indicate unknown statuses
or continued record-level trajectories.
}
\details{
\code{add_arrows()} wraps a new \code{\link[=geom_segment]{geom_segment()}} layer by adding a zero-length
segment at the right end of swimmer lanes. This approach allows users to
specify \code{arrow_neck_length} which can be useful for tracking and visualizing
time in between markers
}
\examples{
\dontrun{
add_arrows(
  data = patient_status,
  mapping = aes(xend = end_time, y = pt_id),
  arrow_neck_length = time_from_today,
  arrow_colour = "forestgreen",
  arrow_fill = "forestgreen"
)
}
}
