% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{add_marker}
\alias{add_marker}
\title{Add markers of interest to level response trajectories}
\usage{
add_marker(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}, either coming from
a parent \code{ggswim()} function, another \code{add_marker()} call, or a new dataframe
prepared for use with \code{ggswim()}.}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.
More information about accepted mapping arguments can be found in \strong{Aesthetics}.}

\item{...}{Other arguments passed to \code{add_marker()}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\value{
A ggswim object
}
\description{
"Markers" are used to specify events of interest along response trajectories
across individual lanes.
}
\section{Aesthetics}{

\code{add_marker()} understands the following aesthetics (required aesthetics are in bold)
when using \code{color}/\code{colour} akin to \code{geom_point()}
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{color}}/\strong{\code{colour}}
\item \code{alpha}
\item \code{group}
\item \code{shape}
\item \code{size}
\item \code{stroke}
}

\code{add_marker()} understands the following aesthetics (required aesthetics are in bold)
when using \code{label} akin to \code{geom_label()}. See "Notes" below for additional
considerations and requirements.
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{color}}/\strong{\code{colour}}
\item \strong{\code{label}} *
\item \code{name} *
\item \code{alpha}
\item \code{angle}
\item \code{family}
\item \code{fontface}
\item \code{group}
\item \code{hjust}
\item \code{lineheight}
\item \code{size}
\item \code{vjust}
}

\strong{Note}: \code{add_marker()} \strong{does not} support mapping using \code{fill}.

\strong{Note}: If using \code{label}, a secondary \code{color} reference is required to make
the label appear in the color layer of the legend.

\strong{Note}: If using a dynamic \code{color} call, a specifying \code{name} is required for
aesthetic mapping to render the legend correctly.
}

