% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{add_marker}
\alias{add_marker}
\title{Add markers of interest to level response trajectories}
\usage{
add_marker(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}, either coming from
a parent \code{ggswim()} function, another \code{add_marker()} call, or a new dataframe
prepared for use with \code{ggswim()}.}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{...}{Other arguments passed to \code{geom_point}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\value{
A ggswim object
}
\description{
"Markers" are used to specify events of interest along response trajectories
across individual lanes.
}
