% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{add_marker}
\alias{add_marker}
\title{Add markers of interest to level response trajectories}
\usage{
add_marker(mapping = aes(), data = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{a dataframe prepared for use with \code{ggswim()}, either coming from
a parent \code{ggswim()} function, another \code{add_marker()} call, or a new dataframe
prepared for use with \code{ggswim()}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
A ggswim object
}
\description{
"Markers" are used to specify events of interest along response trajectories
across individual lanes. \code{add_marker()} wraps \code{geom_point()} or \code{geom_label()}
depending on the users designation of \code{color}/\code{colour} and \code{label} arguments.

See \strong{Aesthetics} for more information.
}
\section{Aesthetics}{

\code{add_marker()} understands the following aesthetics (required aesthetics are in bold)
when using \code{color}/\code{colour} similar to \code{geom_point()}
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{color}}/\strong{\code{colour}}
\item \code{name} *
\item \code{alpha}
\item \code{group}
\item \code{shape}
\item \code{size}
\item \code{stroke}
}

\code{add_marker()} understands the following aesthetics (required aesthetics are in bold)
when using \code{label} similar to \code{geom_label()}. See "Notes" below for additional
considerations and requirements.
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{label_vals}
\item \strong{\code{label_names}} *
\item \code{alpha}
\item \code{angle}
\item \code{family}
\item \code{fontface}
\item \code{group}
\item \code{hjust}
\item \code{lineheight}
\item \code{size}
\item \code{vjust}
}

\strong{Notes}:
\itemize{
\item \code{add_marker()} \strong{does not} support mapping using \code{fill}.
\item If using a static/non-mapping \code{color} specifier, a mapping \code{name} is required
for aesthetic mapping to render the legend correctly.
\item If using labels, both \code{label_vals} and \code{label_names} are required for
proper legend population. At minimum, \code{label_vals} is needed for data
display. These are unique parameter options for \code{aes()} to ggswim.
}
}

\examples{

# markers with points and aesthetic mapping params
add_marker(
  data = infusion_events,
  mapping = aes(
    x = delta_t0,
    y = pt_id,
    color = infusion_type,
    shape = end_study_name
  ),
  size = 5
)

# markers with points and static params

initial_infusions <- infusion_events[infusion_events$delta_t0 == 0, ]

add_marker(
  data = initial_infusions,
  mapping = aes(
    x = delta_t0,
    y = pt_id,
    name = "Initial"
  ),
  color = "red",
  size = 5
)

# markers with labels
add_marker(
  data = end_study_events,
  mapping = aes(y = pt_id, x = delta_t0_months,
      label_names = end_study_name,
      label_vals = end_study_label
  ),
  label.size = NA, fill = NA, size = 5
)
}
