% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{add_marker}
\alias{add_marker}
\title{add_marker title}
\usage{
add_marker(ggswim_obj, data, id, time, ...)
}
\arguments{
\item{ggswim_obj}{a ggswim_obj}

\item{data}{data where markers reside}

\item{id}{the id variable associated with the marker}

\item{time}{the timestamp location for where to place the markers}

\item{...}{Other arguments passed on to layer(). These are often aesthetics,
used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\description{
A short description...
}
