% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_legend.R
\name{bind_layer_data}
\alias{bind_layer_data}
\title{Bind layer dataframes for legend}
\usage{
bind_layer_data(ggswim_obj, layer_indices, layer_data, static_colours = NULL)
}
\arguments{
\item{ggswim_obj}{description}

\item{layer_indices}{description}

\item{layer_data}{description}

\item{static_colours}{description}
}
\value{
A dataframe
}
\description{
Internal helper function that returns layer data from \code{get_layer_data()}
as a bound dataframe to help with legend guide definitions.
}
\keyword{internal}
