% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{bootstrap}
\alias{bootstrap}
\title{Retrieve Bootstrap Unicode}
\usage{
bootstrap(aliases)
}
\arguments{
\item{aliases}{A string or vector of strings to retrieve Unicode values for.}
}
\value{
A named character vector of Unicode values corresponding to the provided aliases.
If an alias is not found, its value will be \code{NA}.
}
\description{
Convert Bootstrap alias strings to their corresponding Unicode format.

All \code{aliases} should be prepended with "bs".
}
\examples{
bootstrap(c("bs-car-front", "bs-heart"))
}
