% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{build_ggswim}
\alias{build_ggswim}
\title{Apply ggswim fixes and display updates}
\usage{
build_ggswim(ggswim_obj)
}
\arguments{
\item{ggswim_obj}{A ggswim object}
}
\value{
A ggswim object
}
\description{
This function seeks to correct the output of manual overrides introduced by
\code{\link[=add_marker]{add_marker()}} depending on the combination of layer types the user provides.

It is to be run automatically in the background as a print method via
\code{\link[=print.ggswim_obj]{print.ggswim_obj()}}.
}
\details{
In its current state, \code{build_ggswim()} can only work with a pre-rendered
ggswim plot object, therefore it cannot be added to the \code{+} operator chain.

\code{build_ggswim()} makes use of \code{\link[ggplot2:guides]{ggplot2::guides()}} to dynamically override
displays in the layers of the ggswim legend. It also applies a call to
\code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}} in applicable cases where a user calls out
a fixed \code{color}/\code{colour} argument in addition to the required \code{color}
mapping aesthetic (handled by arg: \code{name}).
}
\examples{
ggswim_obj <- ggswim(
  data = patient_data,
  mapping = aes(
    x = start_time,
    xend = end_time,
    y = pt_id,
    color = disease_assessment
  ), linewidth = 5
)
build_ggswim(ggswim_obj)
}
