% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_marker_label_aes}
\alias{check_marker_label_aes}
\title{check add_markers for label and color arguments}
\usage{
check_marker_label_aes(mapping)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.}
}
\description{
While not required, the legend will not display with label icons if users don't
supply \code{color} or \code{colour} to \code{aes()} along with \code{label}. Since this doesn't
break the app, a warning should be thrown.
}
\keyword{internal}
