% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_supported_position_args}
\alias{check_supported_position_args}
\title{check for unsupported position args}
\usage{
check_supported_position_args(position, parent_func)
}
\arguments{
\item{position}{Position adjustment. ggswim accepts "identity".  Default "identity".}

\item{parent_func}{The function in which this is being called, to be
referenced in the message output}
}
\description{
\code{ggswim()} accepts \code{position} argument \code{identity} only.
Others such as \code{dodge} and \code{jitter} are not supported.
}
\keyword{internal}
