% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{fontawesome}
\alias{fontawesome}
\title{Retrieve FontAwesome Unicode}
\usage{
fontawesome(aliases, type = "solid")
}
\arguments{
\item{aliases}{A string or vector of strings to retrieve Unicode values for.}

\item{type}{A character string denoting which FontAwesome library subset to search.
One of "solid", "regular", or "brands". Default is "solid".}
}
\value{
A named character vector of Unicode values corresponding to the provided aliases.
If an alias is not found, its value will be \code{NA}.
}
\description{
Convert FontAwesome alias strings to their corresponding Unicode format.

All \code{aliases} should be prepended with "fa".
}
\details{
When using \link{fontawesome} outputs with \link{geom_swim_marker}, the following options
are available for the \code{family} argument:
\itemize{
\item "FontAwesome-Brands"
\item "FontAwesome-Regular"
\item "FontAwesome-Solid"
}
}
\examples{
fontawesome(c("fa-car", "fa-user"))
}
