% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{fontawesome}
\alias{fontawesome}
\title{Retrieve fontawesome unicode}
\usage{
fontawesome(aliases)
}
\arguments{
\item{aliases}{A string or vector of strings to retrieve}
}
\value{
Unicode text
}
\description{
When assigning fontawesome icons as glyphs, \code{\link[=fontawesome]{fontawesome()}} should be used to
convert the alias string to the appropriate Unicode format.

All \code{aliases} should be prepended with "fa".
}
\examples{
fontawesome('fa-car')
}
