% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swim_arrow.R
\docType{data}
\name{geom_swim_arrow}
\alias{geom_swim_arrow}
\alias{GeomSwimArrow}
\title{Add arrows to swimmer plot lanes}
\usage{
geom_swim_arrow(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed",
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{A dataframe prepared for use with \code{\link[=geom_swim_arrow]{geom_swim_arrow()}}. Required.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{arrow_colour}{The colour of the arrow head}

\item{arrow_head_length}{A unit specifying the length of the arrow head
(from tip to base).}

\item{arrow_neck_length}{Value specifying neck length from end of segment
to arrow head base}

\item{arrow_fill}{The fill colour of the arrow head}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head
should be a closed triangle.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Arrows attached to the end of swimmer plot lanes can be used to denote the
continuation of events such as ongoing treatment, implying that the activity
or status extends beyond the plotted period.
}
\details{
Please note that \code{\link[=geom_swim_arrow]{geom_swim_arrow()}} requires a \code{data} argument and does not
inherit data like other functions.
}
\section{Aesthetics}{

\code{\link[=geom_swim_arrow]{geom_swim_arrow()}} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{y}}
\item \strong{xend}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
}

\code{\link[=geom_swim_arrow]{geom_swim_arrow()}} is a wrapper for \code{\link[=geom_segment]{geom_segment()}} and can support much of the same
functionality.
}

\examples{
# Set up data for arrows
arrow_data <- patient_data |>
  dplyr::left_join(
    end_study_events |>
      dplyr::select(pt_id, label),
    by = "pt_id"
  ) |>
  dplyr::select(pt_id, end_time, label) |>
  dplyr::filter(.by = pt_id, end_time == max(end_time)) |>
  dplyr::filter(!is.na(label)) |>
  unique()

geom_swim_arrow(
  data = arrow_data,
  mapping = aes(xend = end_time, y = pt_id),
  linewidth = .1,
  arrow_neck_length = 5,
  arrow_head_length = grid::unit(0.25, "inches"),
  arrow_colour = "slateblue",
  arrow_fill = "cyan"
)

}
\keyword{datasets}
