% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_guide_values}
\alias{get_guide_values}
\title{Create guide values for legend}
\usage{
get_guide_values(
  df,
  gg,
  emoji_or_shape,
  lanes,
  markers,
  groups,
  events,
  group_vals
)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{gg}{A \code{ggplot} object}

\item{emoji_or_shape}{One of "emoji" or "shape", determines whether to use
\code{geom_label()} or \code{geom_point()}}

\item{lanes}{a list of character strings that define the colored line segments
for \code{id}. Colors are supplied by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{markers}{a named list defining marker events on a \code{lane} in either
numeric shape or emoji form}

\item{groups}{additional specifier to indicate groups, optional. Example:
treatment groups or cohorts in a study.}
}
\value{
a list
}
\description{
This function programmatically assigns override values for \code{guide()}.
}
\keyword{internal}
