% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{get_guide_values}
\alias{get_guide_values}
\title{Create guide values for legend}
\usage{
get_guide_values(df, gg, emoji_or_shape, lanes, markers)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{gg}{A \code{ggplot} object}

\item{emoji_or_shape}{One of "emoji" or "shape", determines whether to use
\code{geom_label()} or \code{geom_point()}}

\item{lanes}{Columns that indicate line changes, i.e. color changes
for individual swim lanes.}

\item{markers}{A character vector that will comprise point markers on the
swimmer plot. Optional, default \code{NULL}}
}
\value{
a variable
}
\description{
Programmatically assign and retrieve line and shape/label values to supply to
\code{ggplot2::guide()}.
}
\keyword{internal}
