% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_layer_data}
\alias{get_layer_data}
\title{Build Layer Data}
\usage{
get_layer_data(data, mapping, i = 1L)
}
\arguments{
\item{data}{...}

\item{mapping}{...}

\item{i}{An integer to supply for the layer to retrieve.
If none given, defaults to \code{1L}.}
}
\value{
A dataframe
}
\description{
A short description...
}
\keyword{internal}
