% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim_new.R
\name{get_layer_data}
\alias{get_layer_data}
\title{Build Layer Data}
\usage{
get_layer_data(data, mapping)
}
\arguments{
\item{data}{...}

\item{mapping}{...}
}
\value{
A dataframe
}
\description{
A short description...
}
\keyword{internal}
