% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggswim.R
\name{get_overrides}
\alias{get_overrides}
\title{Get overrides list for \code{guides()}}
\usage{
get_overrides(ref_guide, label_layer_data, point_layer_data)
}
\arguments{
\item{ref_guide}{A reference object from \code{get_guide_data()} to assist with computation}

\item{label_layer_data}{dataframe related to the label layer data for a ggswim_obj}

\item{point_layer_data}{dataframe related to the point layer data for a ggswim_obj}
}
\value{
a list
}
\description{
Creates a list of override definitions to pass to \code{guides()} \code{override.aes}.
}
\keyword{internal}
