% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim-package.R
\docType{package}
\name{ggswim-package}
\alias{ggswim-package}
\alias{_PACKAGE}
\title{ggswim: Create Swimmer Plots with ggplot2}
\description{
Create swimmer plots using the ggplot2 framework.
}
\author{
\strong{Maintainer}: Richard Hanna \email{richardshanna91@gmail.com}

Authors:
\itemize{
  \item Stephan Kadauke \email{kadaukes@chop.edu}
  \item Ezra Porter \email{ezrajporter@gmail.com}
}

}
\keyword{internal}
