% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot swimmer data}
\usage{
ggswim(x, title = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A data object of type \code{swim_tbl} as supplied by \code{streamline()}}

\item{title}{the plot title}

\item{xlab}{the x-axis label}

\item{ylab}{the y-axis label}

\item{...}{arguments passed to ...}
}
\value{
a ggplot2 figure
}
\description{
Plot swimmer data.
}
