% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim_new.R
\name{ggswim}
\alias{ggswim}
\title{ggswim title}
\usage{
ggswim(data, id, time, lane)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time}{the x-axis variable of the swimmer plot, typically a
function of time given in date or numeric format}

\item{lanes}{a column in \code{df} that represents segments of the swimmer plot}
}
\description{
A short description...
}
