% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot swimmer data}
\usage{
ggswim(
  df,
  id,
  time,
  events,
  reference_event,
  emoji_or_shape,
  markers,
  lanes,
  lane_colors = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time}{the x-axis variable of the swimmer plot, typically a
function of time}

\item{events}{the column that will supply definitions for the \code{reference_event},
\code{markers}, and \code{lanes}}

\item{reference_event}{a character string found in \code{events} that establishes
the time-zero reference point for the x-axis}

\item{emoji_or_shape}{One of "emoji" or "shape", determines whether to use
\code{geom_label()} or \code{geom_point()}}

\item{markers}{A character vector that will comprise point markers on the
swimmer plot. Optional, default \code{NULL}}

\item{lanes}{Columns that indicate line changes, i.e. color changes
for individual swim lanes.}

\item{lane_colors}{a vector of character strings to assign to the \code{lanes},
written in the same order. Optional, if left \code{NULL}, default colors applied}

\item{title}{the plot title}

\item{xlab}{the x-axis label}

\item{ylab}{the y-axis label}
}
\value{
a ggplot2 figure
}
\description{
Plot swimmer data.
}
