% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot individual level response trajectories}
\usage{
ggswim(
  df,
  id,
  time,
  events,
  reference_event,
  markers,
  lanes,
  legend_title = NULL
)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time}{the x-axis variable of the swimmer plot, typically a
function of time given in date or numeric format}

\item{events}{the column that will supply data for the \code{reference_event},
\code{markers}, and \code{lanes} arguments}

\item{reference_event}{a character string found in \code{events} that establishes
the time-zero reference point for the x-axis for use when \code{time} is given in
a date/date-time format rather than in an integer/numeric format}

\item{markers}{a named list defining marker events on a \code{lane} in either
standard numeric ggplot 2 shapes, emoji, or unicode form (ex: "\U1F464")}

\item{lanes}{a list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{legend_title}{the title of the legend}
}
\value{
a ggplot2 figure
}
\description{
Visualize individual record response trajectories over time using a swimmer plot.
}
\details{
A swimmer plot is a data visualization used to display individual
subject data over time. It shows events or outcomes as points along a
horizontal line for each subject, allowing easy comparison and pattern
identification.
}
