% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot individual level response trajectories}
\usage{
ggswim(data, mapping = aes(), ...)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.
More information about accepted mapping arguments can be found in \strong{Aesthetics}.}

\item{...}{Other arguments passed to \code{ggswim()}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\description{
Visualize individual record response trajectories over time using a swimmer plot.
}
\details{
A swimmer plot is a data visualization used to display individual
subject data over time. It shows events or outcomes as points along a
horizontal line for each subject, allowing easy comparison and pattern
identification.
}
\section{Aesthetics}{

\code{ggswim()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item fill
\item group
\item linetype
\item linewidth
}

\strong{Note}: \code{ggswim()} \strong{does not} support mapping using \code{color}/\code{colour}.
}

