% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Create swimmer survival plots}
\usage{
ggswim(
  data,
  mapping = aes(),
  position = "identity",
  arrow = NULL,
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed",
  ...
)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{\link[=ggswim]{ggswim()}}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}

\item{arrow}{A column indicating what swim lanes should have arrows applied.
The column must be a logical data type (T/F).}

\item{arrow_colour}{Border/line color to use for the arrow. Default "black".}

\item{arrow_head_length}{A unit specifying the length of the arrow head (from tip to base).
Must be a ggplot2 \code{unit()} object. Default \code{ggplot2::unit(0.25, "inches")}.}

\item{arrow_neck_length}{The length of the neck of the arrow from the end of a
swim lane to the base of the arrow head. Either an integer value or a column
specifier. The default, \code{NULL}, sets a value proportional to the max lane value.}

\item{arrow_fill}{Fill color/colour to use for the arrow head (if closed). Default \code{NULL}.}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head should
be a closed triangle. Default "closed."}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Use ggplot2 architecture to create a swimmer plot showing subject survival
timelines.
}
\details{
A swimmer plot is a data visualization used to display individual
subject data over time. It shows events or outcomes as points along a
horizontal line for each subject, allowing easy comparison and pattern
identification.
}
\section{Aesthetics}{

\code{ggswim()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{xend \emph{or} yend}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
}

\code{ggswim()} is a wrapper for \code{\link[=geom_segment]{geom_segment()}} and can support much of the same
functionality.
}

\section{Arrows}{

Arrows can be specified in \code{ggswim()} as well as via the separate function,
\code{\link[=add_arrows]{add_arrows()}}.
}

\examples{
# Simple ggswim call
ggswim(
  data = patient_data,
  mapping = aes(
    x = start_time,
    xend = end_time,
    y = pt_id,
    color = disease_assessment
  )
)

# ggswim call with arrows
ggswim(
  data = patient_data,
  mapping = aes(
    x = start_time,
    xend = end_time,
    y = pt_id,
    color = disease_assessment
  ),
  arrow = status,
  arrow_fill = "forestgreen",
  arrow_colour = "cyan",
  arrow_head_length = ggplot2::unit(.25, "inches"),
  arrow_neck_length = status_length
)
}
