% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot individual level response trajectories}
\usage{
ggswim(
  data,
  mapping = aes(),
  position = "identity",
  arrow = NULL,
  arrow_colour = "black",
  arrow_length = unit(0.25, "inches"),
  arrow_fill = NULL,
  arrow_type = "closed",
  ...
)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply mapping if there is no plot mapping.
More information about accepted mapping arguments can be found in \strong{Aesthetics}.}

\item{position}{Position adjustment. ggswim accepts either "stack", or "identity"
depending on the use case. Default "identity".}

\item{arrow}{A column indicating what swim lanes should have arrows applied.
The column must be a logical data type (T/F).}

\item{arrow_colour}{Border/line color to use for the arrow. Default "black".}

\item{arrow_length}{A unit specifying the length of the arrow head (from tip to base).
Must be a ggplot2 \code{unit()} object. Default \code{ggplot2::unit(0.25, "inches")}.}

\item{arrow_fill}{Fill color/colour to use for the arrow head (if closed). Default \code{NULL}.}

\item{arrow_type}{One of "open" or "closed" indicating whether the arrow head should
be a closed triangle. Default "closed."}

\item{...}{Other arguments passed to \code{ggswim()}, often aesthetic fixed values,
i.e. \code{color = "red"} or \code{size = 3}.}
}
\description{
Visualize individual record response trajectories over time using a swimmer plot.
}
\details{
A swimmer plot is a data visualization used to display individual
subject data over time. It shows events or outcomes as points along a
horizontal line for each subject, allowing easy comparison and pattern
identification.
}
\section{Aesthetics}{

\code{ggswim()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
}

\strong{Note}: \code{ggswim()} \strong{does not} support mapping using \code{color}/\code{colour}.
}

\examples{
ggswim(
  data = patient_status,
  mapping = aes(
    x = value,
    y = subject_id,
    fill = cohort
  ),
  position = "identity"
)
}
