% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswim.R
\name{ggswim}
\alias{ggswim}
\title{Plot swimmer data}
\usage{
ggswim(
  df,
  id,
  time,
  events,
  reference_event,
  emoji_or_shape,
  markers,
  lanes,
  legend_title = NULL
)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{ggswim()}}

\item{id}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time}{the x-axis variable of the swimmer plot, typically a
function of time}

\item{events}{the column that will supply definitions for the \code{reference_event},
\code{markers}, and \code{lanes}}

\item{reference_event}{a character string found in \code{events} that establishes
the time-zero reference point for the x-axis for use when \code{time} is given in
a date/date-time format rather than in an integer/numeric format}

\item{emoji_or_shape}{One of "emoji" or "shape", determines whether to use
\code{geom_label()} or \code{geom_point()}}

\item{markers}{a named list defining marker events on a \code{lane} in either
numeric shape or emoji form}

\item{lanes}{a list of character strings that line segments for \code{id}. Colors
are also supplied here by setting list elements equal to hex or named colors.
In the absence of colors, default \code{ggplot2} colors will be supplied.}

\item{legend_title}{the title of the legend}
}
\value{
a ggplot2 figure
}
\description{
Plot swimmer data.
}
\details{
Lengths of objects supplied to the following parameters
must be equal to how many exist in the \code{df} supplied:
\itemize{
\item \code{markers}
\item \code{lane_colors}
}

Note: \code{lane_colors} must have the same number of elements as \code{lanes}, even if
not all \code{lanes} are present in \code{df}.
}
