% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griddraw.R
\name{grid.draw.ggswim_obj}
\alias{grid.draw.ggswim_obj}
\title{Draw ggswim object}
\usage{
\method{grid.draw}{ggswim_obj}(x, recording = TRUE)
}
\arguments{
\item{x}{an object of class 'ggswim_obj'}

\item{recording}{A logical value to indicate whether the drawing
    operation should be recorded on the Grid display list.}
}
\value{
None
}
\description{
Draw ggswim object
}
\examples{
# TODO: add example
}
