% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griddraw.R
\name{grid.draw.ggswim_obj}
\alias{grid.draw.ggswim_obj}
\title{Draw ggswim object}
\usage{
\method{grid.draw}{ggswim_obj}(x, recording = TRUE)
}
\arguments{
\item{x}{an object of class 'ggswim_obj'}

\item{recording}{A logical value to indicate whether the drawing
    operation should be recorded on the Grid display list.}
}
\value{
None
}
\description{
\code{grid::grid.draw()} methods for objects of classes 'ggswim_obj' and 'ggcuminc'.
These are implemented to allow users to directly call \code{ggplot2::ggsave()}
on 'ggswim' figures.
}
\examples{
p <- ggplot2::ggplot(data = patient_data) +
 geom_swim_lane(
  mapping = aes(x = start_time,
                y = pt_id,
                xend = end_time,
                color = disease_assessment)
                )
grid.draw(p)
}
\keyword{internal}
