% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_override}
\alias{insert_override}
\title{Insert override layer element}
\usage{
insert_override(data, layer_obj, current_layer, mapping, ignore_mapping)
}
\arguments{
\item{data}{a dataframe prepared for use with \code{ggswim()}}

\item{layer_obj}{a ggswim layer object}

\item{current_layer}{An integer value corresponding to the current working layer}

\item{mapping}{description}

\item{ignore_mapping}{A vector of characters for mapping values to ignore.
Will typically match "required" aes elements from ggplot geom functions}
}
\value{
A ggplot layer object appended with \code{overrides}
}
\description{
A short description...
}
\keyword{internal}
