% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{load_fonts}
\alias{load_fonts}
\title{Load Select Fonts}
\usage{
load_fonts(
  fonts = c("bootstrap-icons", "fa-brands-400", "fa-regular-400", "fa-solid-900"),
  verbose = TRUE
)
}
\arguments{
\item{fonts}{A character vector of font families to load. Valid options are
\code{"bootstrap-icons"}, \code{"fa-brands-400"}, \code{"fa-regular-400"}, and \code{"fa-solid-900"}.
The default is to load all supported fonts.}

\item{verbose}{Whether or not to display load output messages. Default is \code{TRUE}.}
}
\value{
Invisibly returns the font registry as a tibble.
}
\description{
Load open source fonts from ggswim's GitHub repository to the user's cache
directory.

Font files and their associated licenses are available on the main branch.
}
\details{
Currently the following fonts are supported:
\itemize{
\item FontAwesome
\itemize{
\item Brands
\item Regular
\item Solid
}
\item Bootstrap
}

Users can choose which fonts to load by providing a vector of font family names.
The default is to load all supported fonts.
}
