% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{load_fonts}
\alias{load_fonts}
\title{Load Select Fonts}
\usage{
load_fonts(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether or not to display load output messages. Default \code{TRUE}.}
}
\description{
Load open source fonts from the web into the environment for use with ggswim
functions.
}
\details{
Currently the following fonts are supported:
\itemize{
\item FontAwesome
\itemize{
\item Brands
\item Regular
-Solid
}
}
}
