% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{patient_data}
\alias{patient_data}
\title{Patient data}
\format{
A list with the following tibbles:
\describe{
\item{patient_status}{A 5 column tibble with patient IDs, cohorts, status, and start and end times}
\item{adverse_events}{A 4 column tibble with patient IDs, AE names, and time of event}
\item{medication administration}{A 5 column tibble with patient IDs, labels for
medications, names of medications, and time of event}
}
}
\usage{
patient_data
}
\description{
A sample list dataset containing patient status, adverse event, and
medication distribution information. Suitable for a \code{ggswim()} display.
}
\examples{
patient_data
}
\keyword{datasets}
