% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{patient_data}
\alias{patient_data}
\alias{patient_status}
\alias{adverse_events}
\alias{medication_administration}
\title{Patient Dataset}
\format{
A list containing three tibbles:
\describe{
\item{patient_status}{A tibble containing information about patient status
including subject ID, cohort, status, name, value, and time sorting.}
\item{adverse_events}{A tibble containing information about adverse
events including subject ID, adverse event name, and time of event.}
\item{medication_administration}{A tibble containing information about
medication administration including subject ID, medication name, name,
and time of event.}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\source{
This dataset is for demonstration purposes only.
}
\usage{
data(patient_data)

patient_status

adverse_events

medication_administration
}
\description{
This dataset provides information about patients, their status, adverse events,
and medication administration.
}
\examples{
patient_data

patient_data$patient_status

patient_status

adverse_events

medication_administration
}
\keyword{datasets}
