% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{patient_data}
\alias{patient_data}
\title{Patient Data dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53 rows and 4 columns.
}
\source{
This dataset is for demonstration purposes only.
}
\usage{
data(patient_data)
}
\description{
These datasets provide information about patients, infusion events, and end of
study events. Trial data has been de-identified and randomized for general use.
}
\details{
\code{patient_data} contains pre-formatted time series data related to disease status
markers and status markers that can support arrows. This dataset is most
applicable for use with \code{ggswim()} and \code{add_marker()}.
}
\examples{
patient_data
}
\keyword{datasets}
