% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggswim_obj}
\alias{print.ggswim_obj}
\title{Print ggswim object}
\usage{
\method{print}{ggswim_obj}(x, ...)
}
\arguments{
\item{x}{a ggswim object}

\item{...}{not used}
}
\value{
a ggplot
}
\description{
Print ggswim object
}
\examples{
ggplot2::ggplot(data = patient_data) +
  geom_swim_lane(
    mapping = aes(
      x = start_time,
      y = pt_id,
      xend = end_time,
      color = disease_assessment
    )
  ) |>
  print()
}
\keyword{internal}
