% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggswim_obj}
\alias{print.ggswim_obj}
\title{Print ggswim object}
\usage{
\method{print}{ggswim_obj}(x, ...)
}
\arguments{
\item{x}{a ggswim object}

\item{...}{not used}
}
\value{
a ggplot
}
\description{
Print ggswim object
}
\examples{
ggswim(
  data = patient_status,
  mapping = aes(
    x = value,
    y = subject_id,
    fill = cohort
  )
) |>
  print()
}
\keyword{internal}
