% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-ggnewscale.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{new_scale_color}
\alias{new_scale_colour}
\alias{grid.draw}
\alias{ggsave}
\alias{\%+\%}
\alias{aes}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggnewscale}{\code{\link[ggnewscale:new_scale]{new_scale_color}}, \code{\link[ggnewscale:new_scale]{new_scale_colour}}}

  \item{ggplot2}{\code{\link[ggplot2:gg-add]{\%+\%}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggsave}}}

  \item{grid}{\code{\link[grid]{grid.draw}}}
}}

