% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_arrow.R
\name{scale_arrow_discrete}
\alias{scale_arrow_discrete}
\title{Position scales for discrete arrow data}
\usage{
scale_arrow_discrete(
  colours = NULL,
  fills = NULL,
  types = NULL,
  limits = NULL,
  ...
)
}
\arguments{
\item{colours}{Arrow outline colours passed to the arrow layer.}

\item{fills}{Arrow fill colours passed to the arrow layer for closed arrows.}

\item{types}{Arrow head types passed to the arrow layer. One of \code{"open"} or
\code{"closed"}.}

\item{limits}{Arrow labels passed to the arrow scale.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale}}
  \describe{
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} includes the levels in the factor. Please note that to display
every level in a legend, the layer should use \code{show.legend = TRUE}.}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{na.value}}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
    \item{\code{aesthetics}}{The names of the aesthetics that this scale works with.}
    \item{\code{minor_breaks}}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (none for discrete, one minor break
between each major break for continuous)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks. Also
accepts rlang \link[rlang:as_function]{lambda} function notation. When
the function has two arguments, it will be given the limits and major
break positions.
}}
    \item{\code{labels}}{One of the options below. Please note that when \code{labels} is a
vector, it is highly recommended to also set the \code{breaks} argument as a
vector to protect against unintended mismatches.
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{call}}{The \code{call} used to construct the scale for reporting messages.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}
}
\description{
\code{\link[=scale_arrow_discrete]{scale_arrow_discrete()}} is used to set discrete arrow aesthetics for swimmer
plot arrows.
}
\examples{
\dontrun{
ggplot2::ggplot() +
  geom_swim_arrow(
    data = arrow_data,
    aes(xend = end_time, y = pt_id, arrow = label)
  ) +
  scale_arrow_discrete(
    colours = "black",
    fills = "black",
    types = "closed",
    limits = "Continuation"
  )
}

}
