% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_arrow.R
\name{scale_arrow_discrete}
\alias{scale_arrow_discrete}
\title{Position scales for discrete arrow data}
\usage{
scale_arrow_discrete(
  colours = NULL,
  fills = NULL,
  types = NULL,
  limits = NULL,
  ...
)
}
\arguments{
\item{colours}{Arrow outline colours passed to the arrow layer.}

\item{fills}{Arrow fill colours passed to the arrow layer for closed arrows.}

\item{types}{Arrow head types passed to the arrow layer. One of \code{"open"} or
\code{"closed"}.}

\item{limits}{Arrow labels passed to the arrow scale.}
}
\description{
\code{\link[=scale_arrow_discrete]{scale_arrow_discrete()}} is used to set discrete arrow aesthetics for swimmer
plot arrows.
}
\examples{
\dontrun{
ggplot2::ggplot() +
  geom_swim_arrow(
    data = arrow_data,
    aes(xend = end_time, y = pt_id, arrow = label)
  ) +
  scale_arrow_discrete(
    colours = "black",
    fills = "black",
    types = "closed",
    limits = "Continuation"
  )
}

}
