% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{search_bootstrap}
\alias{search_bootstrap}
\title{Search for Bootstrap aliases to include in ggswim}
\usage{
search_bootstrap(str = "", approximate = FALSE)
}
\arguments{
\item{str}{A character string alias to search the icon data available.
If left empty, the default "", returns all available aliases.}

\item{approximate}{Use approximate or exact matching, TRUE/FALSE. Default \code{FALSE}.}
}
\value{
Matching aliases from the available Bootstrap data
}
\description{
Check strings against the available aliases for Bootstrap icons.
}
\examples{
search_bootstrap("bs-car-front")

}
