% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{search_fontawesome}
\alias{search_fontawesome}
\title{Search for fontawesome aliases to include in ggswim}
\usage{
search_fontawesome(str = "", approximate = FALSE)
}
\arguments{
\item{str}{A character string alias to search the fontawesome data available.
If left empty, the default, will return all possibilities.}

\item{approximate}{Use approximate or exact matching, TRUE/FALSE. Default \code{FALSE}.}
}
\value{
Matching aliases from the available fontawesome data
}
\description{
Check strings against the available aliases for fontawesome icons.
}
\examples{
search_fontawesome("fa-car")

}
