% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.R
\name{streamline}
\alias{streamline}
\title{Streamline data for ggswim plotting}
\usage{
streamline(df, subject_var, time_var, markers, class_status = NULL)
}
\arguments{
\item{df}{a dataframe prepared for use with \code{streamline()}}

\item{subject_var}{the y-axis variable of a swimmer plot, typically a unique
subject or record identification column}

\item{time_var}{the the x-axis variable of the swimmer plot, typically a
function of time}

\item{markers}{a vector of columns that will comprise point markers on the
swimmer plot}

\item{class_status}{Columns that indicate line changes, i.e. color changes
for individual swim lanes. Optional, default \code{NULL}.}
}
\value{
a swim_tbl object
}
\description{
Prepare a dataset for use with \code{ggswim()} and ensure proper elements are made
available.
}
\details{
Requirements for a dataset include an identifiable primary key column,
a time column, and any number of binary indicator columns.
}
