% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggswim.R
\name{theme_ggswim}
\alias{theme_ggswim}
\title{Apply custom theme styling for ggplot2 plots}
\usage{
theme_ggswim(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{The base font size to use for the plot elements. Default is 12.}

\item{base_family}{The base font family to use for the plot elements. Default is "".}
}
\value{
A ggplot2 theme object.
}
\description{
This function applies custom styling to various elements of ggplot2 plots,
including title, subtitle, caption, axis text, axis titles, and legend text
and legend titles.
}
\details{
This function builds upon the \code{theme_minimal()} function in ggplot2 but overrides
specific theme elements to provide a customized look and feel for plots.
}
\examples{
p <- ggplot2::ggplot(data = patient_data) +
  geom_swim_lane(
    mapping = aes(
      x = start_time,
      y = pt_id,
      xend = end_time,
      color = disease_assessment
    )
  ) +
  ggplot2::scale_color_brewer(name = "Lanes", palette = "Set1") +
  new_scale_color() +
  geom_swim_point(
    data = infusion_events,
    mapping = aes(x = time_from_initial_infusion, y = pt_id, color = infusion_type), # nolint: object_usage_linter
    size = 5
  ) +
  ggplot2::scale_color_manual(name = "Markers", values = c("red", "green"))

p +
  theme_ggswim()
}
