% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggswim.R
\name{theme_ggswim}
\alias{theme_ggswim}
\title{ggswim Theme}
\usage{
theme_ggswim()
}
\value{
a ggplot2 theme
}
\description{
Alternative themes for ggswim plots:
\itemize{
\item \code{theme_ggswim()}: Builds on \code{theme_minimal()} with removal of grid lines and some aesthetic updates.
}
}
\examples{
ggswim(
  data = patient_data,
  mapping = aes(
    x = delta_t0_months,
    y = pt_id,
    fill = disease_assessment
  )
) +
  theme_ggswim()
}
