% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggswim.R
\name{theme_ggswim}
\alias{theme_ggswim}
\title{ggswim Theme}
\usage{
theme_ggswim()
}
\description{
An alternative theme for ggswim plots.
}
\examples{
ggswim(
  data = patient_status,
  mapping = aes(
    x = value,
    y = subject_id,
    fill = cohort
  )
) +
  theme_ggswim()
}
